+++ src/test/groovy/ru/vyarus/dropwizard/guice/debug/JerseyReportTest.groovy
@@ -1,0 +1,40 @@
+package ru.vyarus.dropwizard.guice.debug
+
+import io.dropwizard.Application
+import io.dropwizard.Configuration
+import io.dropwizard.setup.Bootstrap
+import io.dropwizard.setup.Environment
+import ru.vyarus.dropwizard.guice.GuiceBundle
+import ru.vyarus.dropwizard.guice.test.spock.UseDropwizardApp
+import spock.lang.Specification
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 26.10.2019
+ */
+@UseDropwizardApp(App)
+class JerseyReportTest extends Specification {
+
+    def "Check report correcness"() {
+
+        // actual reporting checked manually (test used for reporting configuration)
+
+        expect: "app does not fail"
+        true
+    }
+
+    static class App extends Application<Configuration> {
+
+        @Override
+        void initialize(Bootstrap<Configuration> bootstrap) {
+            bootstrap.addBundle(GuiceBundle.builder()
+                    .enableAutoConfig("ru.vyarus.dropwizard.guice.cases.hkscope.support")
+                    .printJerseyConfig()
+                    .build())
+        }
+
+        @Override
+        void run(Configuration configuration, Environment environment) throws Exception {
+        }
+    }
+}